+++ fail/fail_test.go
@@ -1,1 +1,7 @@
 package fail
+
+import "testing"
+
+func TestFail(t *testing.T) {
+	// TODO: MissingParams
+}