+++ model/schema_get_test.go
@@ -1,0 +1,21 @@
+package model
+
+import "testing"
+
+func TestGet(t *testing.T) {
+	// TODO: Schema.Get
+	// TODO: Schema.GetAt
+	// TODO: Schema.GetOnly
+	// TODO: Schema.GetOnlyAt
+}
+
+func TestGetRows(t *testing.T) {
+	// TODO: Schema.GetRows
+	// TODO: Schema.GetRowsAt
+	// TODO: Schema.GetRowsOnly
+	// TODO: Schema.GetRowsOnlyAt
+	// TODO: Schema.GetAllRows
+	// TODO: Schema.GetAllRowsAt
+	// TODO: Schema.GetAllRowsOnly
+	// TODO: Schema.GetAllRowsOnlyAt
+}